Move Orchestration CRD into a separate package to cut dependency
