+++ cmd/crd/BUILD.bazel
@@ -10,7 +10,7 @@ go_library(
         "//pkg/composition/crd:go_default_library",
         "//pkg/formation/crd:go_default_library",
         "//pkg/ops:go_default_library",
-        "//pkg/orchestration:go_default_library",
+        "//pkg/orchestration/crd:go_default_library",
         "//vendor/github.com/atlassian/smith/pkg/resources:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
