+++ pkg/orchestration/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "admission.go",
-        "client.go",
         "controller.go",
     ],
     importpath = "github.com/atlassian/voyager/pkg/orchestration",
@@ -12,8 +11,6 @@ go_library(
     deps = [
         "//:go_default_library",
         "//pkg/admission:go_default_library",
-        "//pkg/api/schema:go_default_library",
-        "//pkg/apis/orchestration:go_default_library",
         "//pkg/apis/orchestration/v1:go_default_library",
         "//pkg/k8s:go_default_library",
         "//pkg/k8s/updater:go_default_library",
@@ -30,7 +27,6 @@ go_library(
         "//vendor/go.uber.org/zap:go_default_library",
         "//vendor/k8s.io/api/admission/v1beta1:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
@@ -42,10 +38,7 @@ go_library(
 go_test(
     name = "go_default_test",
     size = "small",
-    srcs = [
-        "client_test.go",
-        "controller_test.go",
-    ],
+    srcs = ["controller_test.go"],
     data = glob(["testdata/**"]),
     embed = [":go_default_library"],
     race = "on",
@@ -55,11 +48,6 @@ go_test(
         "//pkg/orchestration/client/fake:go_default_library",
         "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/atlassian/smith/pkg/apis/smith/v1:go_default_library",
-        "//vendor/github.com/ghodss/yaml:go_default_library",
-        "//vendor/github.com/go-openapi/errors:go_default_library",
-        "//vendor/github.com/go-openapi/spec:go_default_library",
-        "//vendor/github.com/go-openapi/strfmt:go_default_library",
-        "//vendor/github.com/go-openapi/validate:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/require:go_default_library",
