+++ pkg/orchestration/crd/testdata/test_crd_does_not_require_resources.yml
